requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
